requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.30.0
